+++ discovery-provider/src/monitors/monitor_names.py
@@ -21,3 +21,4 @@ transferred_bytes_per_sec = "transferred_bytes_per_sec"
 redis_num_keys = "redis_num_keys"
 redis_used_memory = "redis_used_memory"
 redis_total_memory = "redis_total_memory"
+celery_tasks = "celery_tasks"
