+++ discovery-provider/src/queries/health_check.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from flask import Blueprint, request
 from src.api_helpers import success_response
 from src.queries.get_alembic_version import get_alembic_version
-from src.queries.get_celery_tasks import get_tasks
+from src.queries.get_celery_tasks import get_celery_tasks
 from src.queries.get_db_seed_restore_status import get_db_seed_restore_status
 from src.queries.get_health import get_health, get_latest_ipld_indexed_block
 from src.queries.get_latest_play import get_latest_play
@@ -137,8 +137,8 @@ def ip_check():
 
 @bp.route("/celery_tasks_check", methods=["GET"])
 def celery_tasks_check():
-    tasks = get_tasks()
-    return success_response({"celery_tasks": tasks}, sign_response=False)
+    tasks = get_celery_tasks()
+    return success_response(tasks, sign_response=False)
 
 
 @bp.route("/db_seed_restore_check", methods=["GET"])
